(chore) reduce hook complexity, remove warnings, clean up
