+++ hook/src/lib.rs
@@ -1,6 +1,5 @@
 mod hooks;
 mod patch;
-mod hook;
 mod log;
 mod elf;
